Fixed registration issue
