+++ ethwager/src/server/models/userModel.js
@@ -1,0 +1,18 @@
+const mongoose = require("mongoose");
+
+const UserSchema = new mongoose.Schema({
+  email: {
+    type: String,
+    required: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+});
+
+
+
+const userModel = mongoose.model("User", UserSchema);
+
+module.exports = userModel;
